Add proto format details to AccessLog specification. (#83)
When doing logging to RPC endpoints in proto format we need to know two
config details:

The RPC endpoint. This is a string like the path, but rather than
semantic overloading make this explicit.

Custom headers to include. We'll populate all of the names fields in
the AccessLog proto but users can also specify custom request and
response headers they want to log.