test/api: basic C++ build/link test, minor RLDS fix.
This is useful in Travis CI (and also internally at Google in our CI) to
validate basic build/link of the protos.

This exposed a small issue with API compatibility. Since we don't have
additional package namespaces for each individual API, there was a
conflict between the RateLimit mesage in RLDS and RDS. The quick fix was
to move the RLDS message inside the response object (this is fine as
nobody is using the v2 RLDS yet, open to alternatives including per-xDS
API namespaces).