eds/rds: endpoint subsetting in clusters based on metadata. (#94)
Add metadata to endpoints in EDS.

Allow routes to specify a metadata_match. WeightedClusters can further
specify an additional metadata_match based on weighting.

Log all metadata. It makes sense to have available to the logging
pipeline all metadata used during request processing, not just some
that is explicitly annotated. This then makes available the endpoint's
metadata for log consumption. The tradeoff is that this will increase
log sizes when metadata is plentiful. We may want to consider a way to
filter this in logs in the future.

Fixes #81 and #91.