eds: make UpstreamLocalityStats work with scalar totals.
Previously, we had a complex way of inferring total error requests from
map value aggregation. This PR removes the error maps for now, providing
simple scalar totals instead. This will simplify
https://github.com/lyft/envoy/issues/1286.

We expect to add the error maps as an additional mechanism, orthogonal to
the scalar totals, later on, as standard proto extension.