bootstrap: v1 parity with static config. (#156)
This is a design-level update to bootstrap.proto, that plumbs in the
remaining top-level config from v1. It will probably have some small
changes made beyond this as we implement.

Notable differences to v1 are:

* Static/dynamic resources are clearly delineated at top-level, clusters no longer belong to the ClusterManager object.

* Stats sinks are a repeated list of opaque configs, similar to filter.

* Some simplifications to object types, e.g. RLS no longer specifies type (do we want to preserve the v1 generality here?).

Also renamed RLDS back to RLS, I'll admit that it didn't make sense to
cram it into the xDS namespace, it's really a very distinct service on
the data plane and shouldn't be bundled with the control plane services.