ads: workaround for proto descriptor issue, file reorg and cleanup.
Turns out that files with just service methods don't get loaded into the
descriptor pool automatically in C++. So, needed to have some messages
in ads.proto. Turns out this was a good opportunity to move some of the
messages that were related to discovery out of base.proto.

This shouldn't break the API, since everything is in the envoy.api.v2
packge space.