build: introduce VERSION file
As a source of entire Envoy version number. Will use this file in envoy
main repo as well.

By this change, document version string will be changed in development
version:

  from: "1.6.0-data-plane-api-${GIT_SHA}"
  to: "1.6.0.dev-data-plane-api-${GIT_SHA}"

And in release version, for example, in v1.6.0 release:

  from: "1.6.0-tag-v1.6.0"
  to: "tag-v1.6.0"

The significant change is dropping the first version number string like
"1.6.0-" in the release version document.

Signed-off-by: Taiki Ono <taiks.4559@gmail.com>