Healthcheck filter: compute response based on upstream cluster health
*Description*:
Extend the "non pass through" mode of the HTTP health check filter to allow an
optional map from cluster names to percentages. If specified, each of the named
clusters must have at least the specified percentage of its servers healthy for
the filter to return a 200. (The proxy also must be in a non-draining state, or
else the filter will return a 503 like it did previously.)

Associated envoyproxy/envoy issue: https://github.com/envoyproxy/envoy/issues/2362

Signed-off-by: Brian Pane <bpane@pinterest.com>