lds: Add listener_filter_chain. (#400)
Add listener_filter_chain to Listener, which is a list of individual
listener filters that make up the filter chain for sockets accepted
with the listener. These filters are run before any in the
'filter_chains', and these filters have the opportunity to manipulate
and augment the connection metadata that is used in connection filter
chain matching.  Order matters as the filters are processed
sequentially right after a socket has been accepted by the listener,
and before a connection is created.

As an example the functionality controlled by Listener options
'use_original_dst' and 'use_proxy_proto' can be implemented as
listener filters instead of being hard-wired into Envoy's listener
logic. This makes also extensions (like proxy protocol v2) easier to
implement and deploy.

Signed-off-by: Jarno Rajahalme <jarno@covalent.io>