docs: reorganize for readability (#457)
I went through and reorganized things to make the v2 docs more
human browsable. I also did a few misc cleanups. There is a lot
more to do here which I'm hoping to find a contractor to pay to
work on, but this is a step in the right direction.

Signed-off-by: Matt Klein <mklein@lyft.com>