docs: add stub 1.6.0 release notes (#556)
We are changing the procedure to require contributors to add
release notes directly to this repo as they do changes. It's
too much work for the releaser to transfer them over. I'm
going to do this for 1.6.0 but I'm putting this stub in here
so that any ungoing changes that are going to sneak into 1.6.0
can get added directly.

Signed-off-by: Matt Klein <mklein@lyft.com>