health check: adding unhealthy interval overrides
Three new settings are added to override health check intervals for hosts
that are unhealthy or in transitioning state. This is necessary to achieve
feature parity with HAProxy's [`fastinter` and `downinter`](https://www.haproxy.com/documentation/aloha/7-0/traffic-management/lb-layer7/health-checks/#health-check-parameters).

This is the data plane side of the implementation for the issue https://github.com/envoyproxy/envoy/issues/2499.

Signed-off-by: Marcelo Juchem <juchem@gmail.com>