Implement `allow_missing` for JWT requirement. (#8978)
Description:
Implement `allow_missing` (but not failed) mode for JWT requirement. In this mode, token is not required for JWT filter to pass. However, if a token is presented, it must be valid.

This PR also change the `allow_missing_or_failed` (and the new `allow_missing`) to consider the provider within the container OR-list, instead of using the global one that have all providers. As a result, each `allow_xxx` will examine the interested tokens list independently. This correct the behavior when the `allow_xxx` is used multiple times, e.g in different ORs of an AND-of-OR requirements.

Risk Level: Medium
The PR changes the existing behavior of `allow_missing_or_failed`

Testing:
Unit tests.
Docs Changes: Added
Release Notes: Added

Signed-off-by: Diem Vu <diemvu@google.com>

Mirrored from https://github.com/envoyproxy/envoy @ 74436a6303825e0a6873222efff591ea1001cf87