tracer: ensure x-ray segment field is not empty (#10520)
Initially the name of the segment defaulted to the local cluster/node
name. But since that too can be empty, and this tracer throws and
crashes if the segment is empty, then it's best to make the field
required and not use defaults.

Signed-off-by: Marco Magdy <mmagdy@gmail.com>

Mirrored from https://github.com/envoyproxy/envoy @ e8654dbd4ba3b6cc8c33efbef7bf36d85cd882a0