Docs: replace broken link in api/README.md (#15423)
In the api README a link is given for "API overfiew for users" that
leads to a 404.  It looks to be pointing to a v2_overview page that no
longer exists.  I replaced this link with the current overview at:
https://www.envoyproxy.io/docs/envoy/latest/configuration/overview/overview

Signed-off-by: Zach Mandeville <zz@ii.coop>

Mirrored from https://github.com/envoyproxy/envoy @ eff175c9f2bf316358adfa4ab1a9c41fac7243cf