api: add cluster_specifier_plugin to RouteAction (#16944)
This is a new extension point that will allow for dynamic cluster selection.
The plugin configured by it will be responsible for returning a cluster to
Envoy through its API (to be defined).  Since the cluster returned is dynamic,
a new field is added to Route to list all the clusters it may return.  This
allows for proxy implementations that are not using wildcard CDS queries to
pre-fetch clusters.

This feature may ultimately replace the FilterAction mechanism, as it provides
the same functionality for known use cases, but is simpler to implement and
use.

Signed-off-by: Doug Fawley <dfawley@google.com>

Mirrored from https://github.com/envoyproxy/envoy @ 6dc4092fc161b8fcb5f25a972e4783116692d53c