Create new protobuf for restricting header mutations (#19141)
This protobuf will be used initialy by the ext_proc filter to control
which headers may be changed by an external processing server.

This begins to address #14789 .

If the proto and the location are OK, I'll go on to add a common library to test a proposed header mutation against these rules and then incorporate them into the ext_proc filter.

The eventual result is that, by default, an external processor for ext_proc will be able to modify any header, but there will be controls that an administrator can use when connecting to a processor to control whether that processor is actually allowed to make all possible changes.

Risk Level: Low -- just the proto for now

Signed-off-by: Gregory Brail <gregbrail@google.com>

Mirrored from https://github.com/envoyproxy/envoy @ a3e50ecd88943536b4ce412c172a246dacb6288a